Update index.html
created an unordered list using the unordered list tag open the tag and inserted the content in the list item tag and closed the unordered list tag. Also used the ordered list tag to list items in a given sequence enclosed the items in the list item tag before closing the ordered list tag.